add commands to the extras target to pull the reference and tutorial
* Makefile: add commands to the extras target to pull the reference and
tutorial tarballs from John's moeraki.com. Simplified some of the
directory creation rules. The Makefile is rebuilding every file when one
of the changes, which seems broken, but right now it does the job.

We can move to rsync when it's set up (it's better in dealing with tarball
changes than a wget -c).

* tutorial.src: reordered tutorials and commented out lgs' version for
now.